remove un-needed components
